backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow>=10.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.3
django-filter>=23.0
drf-spectacular>=0.27
python-dotenv>=1.0
dj-database-url>=2.1
whitenoise>=6.6
reportlab>=4.0
celery>=5.3
redis>=5.0
